% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{bin_by_interval}
\alias{bin_by_interval}
\title{Group a data frame into bins}
\usage{
bin_by_interval(.data, col, breaks = NULL)
}
\arguments{
\item{.data}{Data frame to bin}

\item{col}{Column to bin by}

\item{breaks}{Number of bins to create, or a numeric vector of two or more
unique cut points to use. Passed to \code{cut()} to create the cuts. If \code{NULL},
a default number of breaks is chosen based on the number of rows in the
data.}
}
\value{
Grouped data frame, similar to those returned by \code{dplyr::group_by()}.
An additional column \code{.bin} indicates the bin number for each group. Use
\code{dplyr::summarize()} to calculate values within each group, or other dplyr
operations that work on groups.
}
\description{
Groups a data frame (similarly to \code{dplyr::group_by()}) based on the values of
a column, by dividing the numerical range of that column into equal-sized
intervals and collecting all rows in each interval.
}
\examples{
suppressMessages(library(dplyr))
cars |>
  bin_by_interval(speed, breaks = 5) |>
  summarize(mean_speed = mean(speed),
            mean_dist = mean(dist))
}
