% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{binned_residuals}
\alias{binned_residuals}
\title{Obtain binned residuals for a model}
\usage{
binned_residuals(fit, term = NULL, breaks = NULL, ...)
}
\arguments{
\item{fit}{The model to obtain residuals for. This can be a model fit with
\code{lm()} or \code{glm()}, or any model that has a \code{residuals()} method.}

\item{term}{If \code{NULL}, bin the residuals with respect to the fitted values.
If non-\code{NULL}, this should be a string identifying a model term that will
be binned.}

\item{breaks}{Number of bins to create, or a numeric vector of two or more
unique cut points to use. Passed to \code{cut()} to create the cuts. If \code{NULL},
a default number of breaks is chosen based on the number of rows in the
data.}

\item{...}{Additional arguments passed on to \code{residuals()}. The most useful
additional argument is typically \code{type}, to select the type of residuals to
produce (such as standardized residuals or deviance residuals).}
}
\value{
Data frame (tibble) with one row per bin, and the following columns:

\item{bin}{Bin number.}
\item{n}{Number of observations in this bin.}
\item{min, max, mean, sd}{Minimum, maximum, mean, and standard deviation of
either the fitted values of observations in this bin, or of the term if
\code{term} is not \code{NULL}.}
\item{resid.mean}{Mean residual in this bin.}
\item{resid.sd}{Standard deviation of residuals in this bin.}
}
\description{
Construct a data frame by binning the fitted values or predictors of a model
into discrete bins, and calculating the average value of the residuals within
each bin.
}
\details{
In many generalized linear models, the residual plots (Pearson or deviance)
are not useful because the response variable takes on very few possible
values, causing strange patterns in the residuals. For instance, in logistic
regression, plotting the residuals versus covariates usually produces two
curved lines.

If we first bin the data, i.e. divide up the observations into \code{breaks} bins
based on their fitted values, we can calculate the average residual within
each bin. This can be more informative: if a region has 20 observations and
its average residual value is large, this suggests those observations are
collectively poorly fit. By default, the binning is with respect to the
fitted values.
}
\examples{
fit <- lm(mpg ~ cyl + disp + hp, data = mtcars)

binned_residuals(fit, breaks = 5)

binned_residuals(fit, term = "cyl")
}
\references{
Gelman, A. and Hill, J. (2006). Data Analysis Using Regression
and Multilevel/Hierarchical Models. Cambridge University Press.
}
\seealso{
\code{\link[=partial_residuals]{partial_residuals()}} for the related partial residuals;
\code{vignette("logistic-regression-diagnostics")} for examples of use and
interpretation of binned residuals in logistic regression;
\code{\link[=bin_by_interval]{bin_by_interval()}} to bin data and calculate other values in each bin
}
