% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{custom_family}
\alias{custom_family}
\title{Family representing a GLM with custom distribution and link function}
\usage{
custom_family(distribution, inverse_link)
}
\arguments{
\item{distribution}{The distribution of the random component. This should be
in the form of a function taking one argument, the vector of values on the
inverse link scale, and returning a vector of draws from the distribution.}

\item{inverse_link}{The inverse link function.}
}
\value{
A family object representing this family
}
\description{
A GLM is specified by a combination of:
}
\details{
- Random component, i.e. the distribution that Y is drawn from
- Link function relating the mean of the random component to the linear predictor
- Linear predictor

Using `custom_family()` we can specify the random component and link
function, while the linear predictor is set in `population()` when setting up
the population relationships.
}
\examples{
# A zero-inflated Poisson family
rzeroinfpois <- function(ys) {
  n <- length(ys)
  rpois(n, lambda = ys * rbinom(n, 1, prob = 0.4))
}

custom_family(rzeroinfpois, exp)
}
