% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{decrypt}
\alias{decrypt}
\title{Decrypt message giving the location of the true plot in a lineup}
\usage{
decrypt(...)
}
\arguments{
\item{...}{Message to decrypt, specifying the location of the true
diagnostics}
}
\value{
The decrypted message.
}
\description{
Decrypts the message printed by \code{model_lineup()} indicating the location of
the true diagnostics in the lineup.
}
