% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_utils.R
\name{drop_factors}
\alias{drop_factors}
\title{Drop factor columns from a data frame}
\usage{
drop_factors(df)
}
\arguments{
\item{df}{Data frame to process}
}
\value{
Data frame without columns that are factors
}
\description{
Issues messages for the columns dropped.
}
\keyword{internal}
