% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{empirical_link}
\alias{empirical_link}
\title{Empirically estimate response values on the link scale}
\usage{
empirical_link(response, family, na.rm = FALSE)
}
\arguments{
\item{response}{Vector of response variable values.}

\item{family}{Family object representing the response distribution and link
function. Only the link function will be used.}

\item{na.rm}{Should \code{NA} values of the response be stripped? Passed to
\code{mean()} when calculating the mean of the response.}
}
\value{
Mean response value, on the link scale: \eqn{g(\bar Y)}{g(Ybar)},
where \eqn{\bar Y}{Ybar} is the mean of \code{response}.
}
\description{
Calculates the average value of the response variable, and places this on the
link scale. Plotting these against a predictor (by dividing the dataset into
bins) can help assess the choice of link function.
}
\details{
A generalized linear model is one where \eqn{Y} is modeled with some
parametric distribution (such as binomial or Poisson) conditional on \eqn{X},
and

\deqn{g(\mathbb{E}[Y \mid X = x]) = \mu(x)}{g(E[Y | X = x]) = \mu(x)}

where \eqn{g} is a link function.

In exploratory data analysis, we can plot \eqn{X} against \eqn{g(Y)} to
evaluate whether the model is adequate. Typically this is done by binning the
data (for example, by values of \eqn{X}), taking the average of \eqn{Y} in
each bin, and then using \eqn{g} to transform the average.

This can be used to make empirical link plots; see
\code{vignette("other-glm-diagnostics")}.
}
\examples{
suppressMessages(library(dplyr))
suppressMessages(library(ggplot2))

# An empirical link plot for `am`, assuming a binomial with logit link:
mtcars |>
  bin_by_interval(disp, breaks = 5) |>
  summarize(
    mean_disp = mean(disp),
    link = empirical_link(am, binomial())
  ) |>
  ggplot(aes(x = mean_disp, y = link)) +
  geom_point()
}
