% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_utils.R
\name{factor_columns}
\alias{factor_columns}
\title{Check whether each column in a data frame is a factor}
\usage{
factor_columns(df)
}
\arguments{
\item{df}{Data frame to process}
}
\value{
Logical vector with same number of entries as columns in \code{df}. Each
entry is \code{TRUE} if the corresponding column is a factor.
}
\description{
Check whether each column in a data frame is a factor
}
\keyword{internal}
