% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{in_interaction}
\alias{in_interaction}
\title{Determine if a predictor is involved in an interaction}
\usage{
in_interaction(formula, predictor)
}
\arguments{
\item{formula}{Model formula}

\item{predictor}{Predictor to check for, as character vector}
}
\value{
\code{TRUE} if in an interaction, \code{FALSE} otherwise
}
\description{
This presently does not work for any term like \code{poly(x, 3)} or \code{I(x^2)},
since the terms object contains those names, not the names of the underlying
predictors.
}
\keyword{internal}
