% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.R
\name{normalize_family}
\alias{normalize_family}
\title{Accept family arguments in the same way as glm(), convert to standardized
form, and throw comprehensible errors.}
\usage{
normalize_family(family, arg = caller_arg(family))
}
\description{
Accept family arguments in the same way as glm(), convert to standardized
form, and throw comprehensible errors.
}
\keyword{internal}
