% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{population_predictors}
\alias{population_predictors}
\title{Get the predictors of a population}
\usage{
population_predictors(population)
}
\arguments{
\item{population}{Population object}
}
\value{
Named list of predictors
}
\description{
Get the predictors of a population
}
\keyword{internal}
