% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{population_response}
\alias{population_response}
\title{Get the response variables of a population}
\usage{
population_response(population)
}
\arguments{
\item{population}{Population object}
}
\value{
Named list of response variables
}
\description{
Get the response variables of a population
}
\keyword{internal}
