% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{potential_outcomes}
\alias{potential_outcomes}
\title{Define a finite population with potential outcomes
TODO}
\usage{
potential_outcomes()
}
\description{
Define a finite population with potential outcomes
TODO
}
