% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_utils.R
\name{prototype_for}
\alias{prototype_for}
\title{Get a prototype data frame for partial residuals}
\usage{
prototype_for(df, predictor)
}
\arguments{
\item{df}{data frame of predictors}

\item{predictor}{character vector identifying one predictor}
}
\value{
prototype data frame
}
\description{
All predictors, except the one we are calculating partial residuals for, are
set to 0 (or their baseline level, for factors).
}
\keyword{internal}
