% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressinator-package.R
\docType{package}
\name{regressinator-package}
\alias{regressinator}
\alias{regressinator-package}
\title{regressinator: Simulate and Diagnose (Generalized) Linear Models}
\description{
Simulate samples from populations with known covariate distributions, generate response variables according to common linear and generalized linear model families, draw from sampling distributions of regression estimates, and perform visual inference on diagnostics from model fits.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.refsmmat.com/regressinator/}
  \item \url{https://github.com/capnrefsmmat/regressinator}
  \item Report bugs at \url{https://github.com/capnrefsmmat/regressinator/issues}
}

}
\author{
\strong{Maintainer}: Alex Reinhart \email{areinhar@stat.cmu.edu} (\href{https://orcid.org/0000-0002-6658-514X}{ORCID})

}
\keyword{internal}
