% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{response}
\alias{response}
\title{Specify a response variable in terms of predictors}
\usage{
response(expr, family = gaussian(), error_scale = NULL)
}
\arguments{
\item{expr}{An expression, in terms of other predictor or response variables,
giving this predictor's value on the link scale.}

\item{family}{The family of this response variable, e.g. \code{gaussian()} for an
ordinary Gaussian linear relationship.}

\item{error_scale}{Scale factor for errors. Used only for linear families,
such as \code{gaussian()} and \code{ols_with_error()}. Errors drawn while simulating
the response variable will be multiplied by this scale factor. The scale
factor can be a scalar value (such as a fixed standard deviation), or an
expression in terms of the predictors, which will be evaluated when
simulating response data. For generalized linear models, leave as \code{NULL}.}
}
\description{
Response variables are related to predictors (and other response variables)
through a link function and response distribution. First the expression
provided is evaluated using the predictors, to give this response variable's
value on the link scale; then the inverse link function and response
distribution are used to get the response value. See Details for more
information.
}
\details{
Response variables are drawn based on a typical generalized linear model
setup. Let \eqn{Y} represent the response variable and \eqn{X} represent the
predictor variables. We specify that

\deqn{Y \sim \text{SomeDistribution}(g^{-1}(\mu(X)))}{%
      Y ~ SomeDistribution(g^{-1}(\mu(X)))}

where \eqn{\mu(X)} is the expression \code{expr}, and both the distribution and
link function \eqn{g} are specified by the \code{family} provided. For instance,
if the \code{family} is \code{gaussian()}, the distribution is Normal and the link is
the identity function; if the \code{family} is \code{binomial()}, the distribution is
binomial and the link is (by default) the logistic link.
}
