% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sample_x}
\alias{sample_x}
\title{Draw a data frame of X values from the specified population.}
\usage{
sample_x(population, n)
}
\arguments{
\item{population}{Population, as defined by \code{population()}}

\item{n}{Number of observations to draw from the population}
}
\value{
Data frame of \code{n} rows, with columns matching the variables specified
in the population
}
\description{
Draw a data frame of X values from the specified population.
}
