% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sample_y}
\alias{sample_y}
\title{Augment the X values with sampled response values}
\usage{
sample_y(xs)
}
\arguments{
\item{xs}{Sample X values drawn from the population, as obtained from
`sample_x()`}
}
\value{
Data frame of `xs` with additional column named `y`
}
\description{
Given the data frame of `xs`, use the true relationship in the population to
draw random Y values corresponding to those X values.
}
\examples{
# A population with a simple linear relationship
pop <- population(
  0.7 + 2.2 * x1 - 0.2 * x2,
  predictors = list(
      x1 = list(dist = "rnorm", mean = 4, sd = 10),
      x2 = list(dist = "runif", min = 0, max = 10)
  ),
  error_scale = 1.0)

pop |>
  sample_x(10) |>
  sample_y()
}
