% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{augment_quantile}
\alias{augment_quantile}
\title{Augment data with randomized quantile residuals}
\usage{
augment_quantile(x, ...)
}
\arguments{
\item{x}{Fitted model to obtain randomized quantile residuals from}

\item{...}{Additional arguments to pass to \code{broom::augment()}}
}
\value{
Data frame with one row per observation used to fit \code{x}, including a
\code{.quantile.resid} column containing the quantile residuals. See
\code{broom::augment()} and its methods for details of other columns.
}
\description{
Generates a data frame with one row per observation used to train a model,
including the predictors, residuals, and randomized quantile residuals as
additional columns.
}
\details{
Randomized quantile residuals provide more interpretable residuals for
generalized linear models, such as logistic regression. See Dunn and Smyth
(1996) for details, or review the examples provided in \code{vignette("DHARMa", package="DHARMa")}.

Uses \code{broom::augment()} to generate the data frame, then uses the \href{https://cran.r-project.org/package=DHARMa}{DHARMa package} to generate randomized
quantile residuals for the model.
}
\references{
Dunn, Peter K., and Gordon K. Smyth (1996).
"Randomized Quantile Residuals." \emph{Journal of Computational and Graphical
Statistics} 5 (3): 236–44. \doi{10.2307/1390802}
}
