% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{binned_residuals}
\alias{binned_residuals}
\title{Obtained binned residuals for a model}
\usage{
binned_residuals(fit, term = NULL, n_bins = NULL, ...)
}
\arguments{
\item{fit}{The model to obtain residuals for. This can be a model fit with
\code{lm()} or \code{glm()}, or any model that has a \code{residuals()} method.}

\item{term}{If \code{NULL}, bin the residuals with respect to the fitted values.
If non-\code{NULL}, this should be a string identifying a model term that will
be binned.}

\item{n_bins}{The number of discrete bins to use. If \code{NULL}, the number of
bins is chosen heuristically to scale with the number of observations, to
ensure each bin has a reasonable number of observations.}

\item{...}{Additional arguments passed on to \code{residuals()}. The most useful
additional argument is typically \code{type}, to select the type of residuals to
produce (such as standardized residuals or deviance residuals).}
}
\value{
Data frame (tibble) with one row per bin.

\item{.bin}{Bin number.}
\item{.n}{Number of observations in this bin.}
\item{.fitted.min, .fitted.max, .fitted.mean, .fitted.sd}{Minimum, maximum,
mean, and standard deviation of the fitted values of observations in this
bin. If \code{term} is not \code{NULL}, these are instead named after the term, e.g. if
\code{term = "x1"} these are \code{x1.min} and so on.}
\item{.resid.mean}{Mean residual in this bin.}
\item{.resid.sd}{Standard deviation of residuals in this bin.}
}
\description{
Construct a data frame by binning the fitted values or predictors of a model
into discrete bins, and calculating the average value of the residuals within
each bin.
}
\details{
In many generalized linear models, the residual plots (Pearson or deviance)
are not useful because the response variable takes on very few possible
values, causing strange patterns in the residuals. For instance, in logistic
regression, plotting the residuals versus covariates usually produces two
curved lines.

If we first bin the data, i.e. divide up the observations into \code{n_bins} bins
based on their fitted values, we can calculate the average residual within
each bin. This can be more informative: if a region has 20 observations and
its average residual value is large, this suggests those observations are
collectively poorly fit. By default, the binning is with respect to the
fitted values.
}
\examples{
fit <- lm(mpg ~ cyl + disp + hp, data = mtcars)

binned_residuals(fit, n_bins = 5)

binned_residuals(fit, term = "cyl")
}
\references{
Gelman, A. and Hill, J. (2006). Data Analysis Using Regression
and Multilevel/Hierarchical Models. Cambridge University Press.
}
\seealso{
\code{\link[=partial_residuals]{partial_residuals()}} for the related partial residuals;
\code{vignette("logistic-regression-diagnostics")} for examples of use and
interpretation of binned residuals in logistic regression
}
