% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{ols_with_error}
\alias{ols_with_error}
\title{Family representing a linear relationship with non-Gaussian errors}
\usage{
ols_with_error(error, ...)
}
\arguments{
\item{error}{Function that can draw random variables from the non-Gaussian
distribution. For example, \code{rt} draws \emph{t}-distributed random variates. The
function must take an argument \code{n} indicating how many random variates to
draw (as all random generation functions built into R do).}

\item{...}{Further arguments passed to the \code{error} function to draw random
variates, such as to specify degrees of freedom, shape parameters, or other
parameters of the distribution. These arguments are evaluated with the
model data in the environment, so they can be expressions referring to
model data, such as values of the predictors.}
}
\value{
A family object representing this family.
}
\description{
This family can represent any non-Gaussian error, provided random variates
can be drawn by an R function. A family specified this way can be used to
specify a population (via \code{population()}), but can't be used to estimate a
model (such as with \code{glm()}).
}
\examples{
# t-distributed errors with 3 degrees of freedom
ols_with_error(rt, df = 3)

# Cauchy-distributed errors
ols_with_error(rcauchy, scale = 3)
}
