% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{parametric_boot_distribution}
\alias{parametric_boot_distribution}
\title{Simulate the distribution of estimates by parametric bootstrap}
\usage{
parametric_boot_distribution(fit, fn = tidy, nsim = 100, ...)
}
\arguments{
\item{fit}{A model fit to data, such as by \code{lm()} or \code{glm()}, to simulate new
response values from.}

\item{fn}{Function to call on each new model fit to produce a data frame of
estimates. Defaults to \code{broom::tidy()}, which produces a tidy data frame of
coefficients, estimates, standard errors, and hypothesis tests.}

\item{nsim}{Number of total simulations to run.}

\item{...}{Additional arguments passed to \code{fn} each time it is called.}
}
\value{
A data frame (tibble) with columns corresponding to the columns
returned by \code{fn}. The additional column \code{.sample} indicates which fit each
row is from.
}
\description{
Repeatedly simulates new response values by using the fitted model, holding
the covariates fixed, and refits the model to each simulated dataset.
Estimates, confidence intervals, or other quantities are extracted from each
fitted model and returned as a tidy data frame.
}
\details{
Because \code{model_lineup()} uses the S3 generic methods \code{model.frame()},
\code{simulate()}, and \code{update()}, it can be used with any model fit for which
methods are provided. In base R, this includes \code{lm()} and \code{glm()}.
}
\examples{
fit <- lm(dist ~ speed, data = cars)
parametric_boot_distribution(fit, nsim = 5)
}
\seealso{
\code{\link[=model_lineup]{model_lineup()}} to use resampling to aid in regression diagnostics;
\code{\link[=sampling_distribution]{sampling_distribution()}} to simulate draws from the population
distribution, rather than the null
}
