% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{population}
\alias{population}
\title{Define the population generalized regression relationship}
\usage{
population(...)
}
\arguments{
\item{...}{A sequence of named arguments defining predictor and response
variables. These are evaluated in order, so later response variables may
refer to earlier predictor and response variables. All predictors should be
provided first, before any response variables.}
}
\value{
A population object.
}
\description{
Specifies a hypothetical infinite population of cases. Each case has some
predictor variables and one or more response variables. The relationship
between the variables and response variables are defined, as well as the
population marginal distribution of each predictor variable.
}
\examples{
# A population with a simple linear relationship
population(
  x1 = predictor("rnorm", mean = 4, sd = 10),
  x2 = predictor("runif", min = 0, max = 10),
  y = response(0.7 + 2.2 * x1 - 0.2 * x2, error_scale = 1.0)
)

# A binary outcome Y, using a binomial family with logistic link
population(
  x1 = predictor("rnorm", mean = 4, sd = 10),
  x2 = predictor("runif", min = 0, max = 10),
  y = response(0.7 + 2.2 * x1 - 0.2 * x2,
               family = binomial(link = "logit"))
)

# A population with a simple linear relationship and collinearity. Because X
# is bivariate, there will be two predictors, named x1 and x2.
library(mvtnorm)
population(
  x = predictor("rmvnorm", mean = c(0, 1),
                sigma = matrix(c(1, 0.8, 0.8, 1), nrow = 2)),
  y = response(0.7 + 2.2 * x1 - 0.2 * x2, error_scale = 1.0)
)
}
\seealso{
\code{\link[=predictor]{predictor()}} and \code{\link[=response]{response()}} to define the population;
\code{sample_x()} and \code{sample_y()} to draw samples from it
}
